requirements.txt
openai==0.27.0
pandas==2.0.3
# requests==2.31.0
openpyxl==3.1.2
python-dotenv==1.0.1